fix: support longer input plans than the argument line limit
Sometimes terraform plans become so large that they do not fit into the
command line as arguments. As the path ${{ github.workspace }} is
mounted automatically as /github/workspace in the container we can use
that to pass the input as file.

To stay backwards compatible we add support to pass the input using a
file.